refactor(web): split web input model types into themed files
