+++ packages/react-native-enriched-markdown/src/web/input/model/rangeBounds.ts
@@ -1,0 +1,6 @@
+// Positions are UTF-16 code units into the plain-text buffer
+// `start` inclusive, `end` exclusive.
+export interface RangeBounds {
+  start: number;
+  end: number;
+}
